moved pagination logic to external file
